Removed FOpen and curl dependency for video.
Removed Curl FOpen Dependency for Video. Updated overwrite for images. Script will now check if there is an article has a new image in client's feed and update the post's thumnbnail image regardless if it has one already or not.

Removed image_update function. It  added unnecessary calls to the database. Replaced it with a more intuitive function that does more than return a post's picture id meta value. 

Quetioned the need for populate_postmeta function since it is not used.